refactor: Model 디렉토리 추가 및 Model 파일 이동
이동한 파일
- Lotto.js
- WinningLotto.js
- LottoMachine.js
- ResultCalculator.js
