+++ src/Model/Lotto.js
@@ -1,7 +1,7 @@
-import ERROR from './constants/error.js';
-import LOTTO from './constants/lotto.js';
-import CustomError from './errors/CustomError.js';
-import { isPositiveInteger } from './utils/function.js';
+import ERROR from '../constants/error.js';
+import LOTTO from '../constants/lotto.js';
+import CustomError from '../errors/CustomError.js';
+import { isPositiveInteger } from '../utils/function.js';
 
 class Lotto {
   #numbers;
