+++ src/Model/LottoMachine.js
@@ -1,9 +1,9 @@
 import { Random } from '@woowacourse/mission-utils';
 import Lotto from './Lotto.js';
-import ERROR from './constants/error.js';
-import LOTTO from './constants/lotto.js';
-import CustomError from './errors/CustomError.js';
-import { isPositiveInteger } from './utils/function.js';
+import ERROR from '../constants/error.js';
+import LOTTO from '../constants/lotto.js';
+import CustomError from '../errors/CustomError.js';
+import { isPositiveInteger } from '../utils/function.js';
 
 class LottoMachine {
   #money;
